+++ src/controller/homeController.js
@@ -1,12 +1,4 @@
-// Get the client
-import mysql from "mysql2";
-
-// Create the connection to database
-const connection = mysql.createConnection({
-  host: "localhost",
-  user: "root",
-  database: "jwt",
-});
+import userService from "../service/userService";
 
 const handleHelloWorld = (req, res) => {
   return res.render("home.ejs");
@@ -21,15 +13,9 @@ const handleCreateNewUser = (req, res) => {
   let password = req.body.password;
   let username = req.body.username;
 
-  connection.query(
-    "INSERT INTO users (email, password,username) VALUES (?,?,?)",
-    [email, password, username],
-    function (err, results, fields) {
-      if (err) {
-        console.log("Error inserting user: ", err);
-      }
-    }
-  );
+  // userService.createNewUser(email, password, username);
+  userService.getUserList();
+
   return res.send("handle create new user");
 };
