+++ src/views/user.ejs
@@ -17,6 +17,11 @@
     <form method="post" action="/user/create-user">
       <div class="container">
         <div class="row">
+          <div class="pt-3">
+            <h3>
+              Create a new user
+            </h3>
+          </div>
           <div class="mb-3 mt-3 col-6">
             <label for="email" class="form-label">Email address</label>
             <input type="email" class="form-control" id="email" name="email" required />
